requirements.txt
annotated-types==0.7.0
anyio==4.8.0
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.1
docx2pdf==0.1.8
exceptiongroup==1.2.2
fastapi==0.115.8
fire==0.7.0
fonttools==4.56.0
h11==0.14.0
idna==3.10
lxml==5.3.1
numpy==2.2.3
opencv-python-headless==4.11.0.86
pandoc==2.4
pdf2docx==0.5.8
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==11.1.0
plumbum==1.9.0
ply==3.11
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyMuPDF==1.25.3
pypandoc==1.15
pypdfium2==4.30.1
python-docx==1.1.2
python-multipart==0.0.20
pywin32==308
reportlab==4.3.1
sniffio==1.3.1
starlette==0.45.3
termcolor==2.5.0
tqdm==4.67.1
typing_extensions==4.12.2
uvicorn==0.34.0